requirements-notebook.txt
# Requirements for running the Jupyter notebook
tensorflow-cpu==2.15.0
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
scikit-learn==1.3.0
pillow==10.1.0
jupyter==1.0.0
notebook==7.0.6
